perf(session): render tool-call content lazily, extract to components/chat
Extract the tool-call rendering (ToolCallDetails + internal ToolCallValueView)
out of the session route into components/chat/ToolCallDetails.tsx, and gate its
inputs/results (and their parsing) on the expanded state via onToggle — collapsed
tool calls no longer pay the parse cost up front.
